Move methods to files based on type
